Add files via upload
R script developed as part of Week 4 Peer Graded Assignment for Getting and Cleaning Data Course Project. The purpose of the script is to merge the test and train datasets included in the UCI HAR dataset specified in the course project instructions, and ultimately satisfy the marking criteria for the peer graded assignment.